fix: viewModel 추가 by Hah-nna
